// generated/src/app/api-models/cap-cptsu-rpt-wrk-temp.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapCptsuRptWrkTemp {
  seqCapCptsuRptWrkTemp: number;
  seqJobId: number;
  capModelId: string;
  firstContractMonth: Date;
  capEntityCode: string;
  seqCapPoolId: number;
  capMonthRun: Date;
  capMonth: Date;
  applyTo: string;
  capStoplossId: string;
  capTransAmt: number;
  pmpmWitholdAmt: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  adjustmentFlag: string;
}